# api_app/docs/swagger.yaml
definitions:
...
  api_model.BaseResponseAPIModel:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api_model.DetailParkingResponseDto:
    properties:
      address:
        type: string
      city_id:
        type: integer
      close_at_hour:
        type: integer
      close_at_minute:
        type: integer
      district_id:
        type: integer
      full_address:
        type: string
      location:
        $ref: '#/definitions/api_model.PointLocationApiDto'
      open_at_hour:
        type: integer
      open_at_minute:
        type: integer
      owner_name:
        type: string
      owner_phone:
        type: string
      parking_id:
        type: string
      parking_name:
        type: string
      parking_phone:
        type: string
      parking_types:
        items:
          $ref: '#/definitions/api_model.ParkingTypesResponseDto'
        type: array
      status:
        type: integer
      status_display:
        type: string
      ward_id:
        type: integer
    type: object
  api_model.DetailPlaceResponseDto:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
...
  api_model.GetCircleParkingLocationModelResponseDto:
    properties:
      address:
        type: string
      distance:
        type: number
      location:
        $ref: '#/definitions/api_model.PointLocationApiDto'
      name:
        type: string
      parking_id:
        type: integer
      parking_phone:
        type: string
      parking_types:
        items:
          $ref: '#/definitions/api_model.ParkingTypesResponseDto'
        type: array
      public_id:
        type: string
      status:
        type: integer
      status_display:
        type: string
    type: object
  api_model.GetDirectionResponseDto:
    properties:
      points:
        items:
          $ref: '#/definitions/api_model.PointLocationApiDto'
        type: array
    type: object
  api_model.GetPlaceHolderResponseDto:
    properties:
      place:
        type: string
      place_id:
        type: string
    type: object
...
  api_model.ParkingTypesResponseDto:
    properties:
      current_slot:
        type: integer
      id:
        type: integer
      logo:
        type: string
      price:
        type: number
      status:
        type: integer
      status_display:
        type: string
      total_slot:
        type: integer
      type:
        type: integer
      type_display:
        type: string
    type: object
  api_model.PointLocationApiDto:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  api_model.RecommendParkingLocationBaseResponseDto:
    properties:
      parking:
        items:
          $ref: '#/definitions/api_model.RecommendParkingLocationResponseDto'
        type: array
    type: object
  api_model.RecommendParkingLocationResponseDto:
    properties:
      location:
        $ref: '#/definitions/api_model.PointLocationApiDto'
      parking_id:
        type: string
      parking_name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/external/v1/{parking_id}:
    get:
      consumes:
      - application/json
      description: Show  the parking detail information
      parameters:
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  $ref: '#/definitions/api_model.DetailParkingResponseDto'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Parking detail information
      tags:
      - external
  /api/external/v1/circle-location:
    get:
      consumes:
      - application/json
      description: Show the top nearest parking in circle radius of specific location
      parameters:
      - description: lat
        in: query
        name: lat
        type: number
      - description: lng
        in: query
        name: lng
        type: number
      - description: distance
        in: query
        name: distance
        type: integer
      - description: text_search
        in: query
        name: text_search
        type: string
      - description: city
        in: query
        name: city
        type: integer
      - description: district
        in: query
        name: district
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.GetCircleParkingLocationModelResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Top nearest parking in circle radius
      tags:
      - external
  /api/external/v1/directions:
    get:
      consumes:
      - application/json
      description: Show the direction from specific location to parking
      parameters:
      - description: origin
        in: query
        name: origin
        type: string
      - description: parking_id
        in: query
        name: parking_id
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  $ref: '#/definitions/api_model.GetDirectionResponseDto'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Show the direction from specific location to parking
      tags:
      - external
  /api/external/v1/placeholder:
    get:
      consumes:
      - application/json
      description: GetPlaceHolder
      parameters:
      - description: input
        in: query
        name: input
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: radius
        in: query
        name: radius
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.GetPlaceHolderResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: GetPlaceHolder
      tags:
      - external
  /api/external/v1/placeholder/{place_id}/detail:
    get:
      consumes:
      - application/json
      description: GetDetailPlaceHolder
      parameters:
      - description: place_id
        in: path
        name: place_id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.DetailPlaceResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: GetDetailPlaceHolder
      tags:
      - external
  /api/external/v1/recommend:
    get:
      consumes:
      - application/json
      description: Show the recommend parking
      parameters:
      - description: page_index
        in: query
        name: page_index
        type: integer
      - description: page_limit
        in: query
        name: page_limit
        type: integer
      - description: city_id
        in: query
        name: city_id
        type: integer
      - description: lat
        in: query
        name: lat
        type: integer
      - description: lng
        in: query
        name: lng
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.RecommendParkingLocationBaseResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Show the recommend parking
      tags:
      - external
...
swagger: "2.0"